Implement decimal type for financial fields and enhance database schema
- Updated the database schema to use decimal types for financial fields in accounts, entries, transfers, and exchange rates to ensure precision in monetary calculations.
- Refactored related SQL queries and Go structs to accommodate the new decimal type.
- Enhanced mock implementations and tests to reflect changes in data types, ensuring robust testing for account and transfer functionalities.
- Introduced a new migration script for the combined schema and removed outdated migration files to streamline the database setup process.

+++ db/query/exchange_rate/create_exchange_rate.sql
@@ -5,4 +5,4 @@ INSERT INTO exchange_rates (
   rate
 ) VALUES (
   $1, $2, $3
-) RETURNING id, from_currency, to_currency, rate::float8, created_at; 
+) RETURNING id, from_currency, to_currency, rate, created_at;
